backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
click==8.3.0
distro==1.9.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
openai==1.3.7
packaging==25.0
pluggy==1.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
PyYAML==6.0.3
redis==5.0.1
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
google-generativeai==0.3.2
